Criação de banco de dados, entidades e daos

+++ app/src/main/java/com/example/gestorgastofamiliar/fragments/FragmentRegistro.kt
@@ -92,7 +92,7 @@ class FragmentRegistro : Fragment() {
                     concept,
                     date,
                     price,
-                    binding.spUsuario.selectedItem.toString()
+                    binding.spUsuario.selectedItem.toString().toInt()
                 )
 
                 GastosProvider.gastos.add(gasto)
